fix an TargetInvocation error for inherited templates
This patches fixes the issue with inherited templates. Create T4
precompiled template SharedFragment.tt with the following text:

<#@ template language="C#" #>
<#+
protected void SharedText(int n)
{
 #>
   Shared Text <#= n #>
<#+
}
// Insert more methods here if required.
 #>

Then create basic T4 template with the code:

<#@ template language="C#" inherits="BaseT4" #>
<#@ import namespace="ConsoleTest" #>
<#@ assembly name="/pathToProject/ConsoleTest/bin/Release/ConsoleTest.exe" #>

begin 1
   <# SharedText(2); #>
end 1

When the template is saving in Monodevelop it fails with TargetInvocation exception. The same
template works normally in Visual Studio. This patch is fixing the issue