[T4] Fix null ref when using parameterized T4 templates.
This also fixes bug #36093 - [ASP.NET Project Wizard] Global.asax
could not be written.
https://bugzilla.xamarin.com/show_bug.cgi?id=36093

The simplest way to reproduce the exception is to create a T4 template
with the content shown below and then try to save the T4 template:

    <#@ template language="C#" #>
    <#@ parameter type="System.Int32" name="TimesToRepeat" #>

The TemplatingEngine.GenerateIndentedClassCode uses reflection when
running on Mono to initialize internal members of the CodeGenerator
class since the implementation of the CSharpCodeProvider is
incomplete on Mono so the CodeDomProvider CreateCodeFromMember method
cannot be called. The CodeGenerator in Mono 4.2 has changed and no
longer has an InitOutput method that the TemplatingEngine was using
to initialize the internal members. This missing method was causing a
null reference exception.

Now if the InitOutput method is missing the TemplatingEngine will
instead directly initialize the internal fields of the CodeGenerator
class. The CSharpCodeProvider's GenerateCodeFromMember is still
not implemented with Mono 4.2 so reflection is still required.

Also added a unit test for the GenerateIndentedClassCode method.