Fixed relative paths that causes an exception in ResolvePath.
When a relative path was used in an include directive the ResolvePath would try to get the directory name and would return null that caused path combine to throw an exception.  The solution is to set dir to an empty string when the directory name returns null.  This will prevent the exception and allow using the included paths passed in on the command line.