Update RoslynCodeCompiler.cs
EmbeddedText.FromSource is throwing an exception with the SyntaxTree.FilePath is null.  In this case, FilePath is always null, so if the template directives have debug="true" then the line I removed is causing an exception to be thrown stating the filePath parameter is null.  The embeddedTexts variable doesn't seem to be used anyway, so I don't believe it's necessary.  Perhaps left-over debug code?