Land the xunit port from build-targets branch
The build-targets branch contains a lot of fixes that
are unrelated to the WIP build targets work, but the
switch from nunit to xunit makes it very difficult to
backport them, as well as making it difficult to merge
main into build-targets.

Backporting the xunit port first makes it much more
straightforward to cherry-pick those fixes onto main.