Use ConfigureAwait(false) in library code
Continuations within the templating engine should not need to
resume on the synchronization context, and because it supports the
older sync APIs by wrapping the async ones, that could cause it to
deadlock.

Add some basic tests using a sync context that records whether it
was called. In future will enable the analyzer that verifies
this behaviour, but that comes with a bunch of other analyzers that
will need review and suppression or fixes.

Should fix #122 (ProcessTemplate does not work in .NET 6 WinForms application_