Re-enable XUnit AppDomain on Mono
This causes the tests to run in an AppDomain where the BaseDirectory
in the directory containing the tests assembly, else it's the directory
containing the xUnit test runner.

This allows the AppDomain created for executing templates to find the
Mono.TextTemplating.dll assembly, which it requires for remoting.

Hopefully the intermittent Mono hang that caused us to disable the xUnit
AppDomain is fixed....

Fixed the templating AppDomain has the same base directory, and can then  to find Mono.TextTemplating.dll,