Add support for C# 11 and newer versions
If a langversion is explicitly provided, use it verbatim. The
compiler will check validity and emit and error if needed.
Otherwise, use the C# language version known to be supported by
the current runtime, as we may be using csc/Roslyn from a newer
runtime where the "latest" language features depend on new APIs
that aren't available on the current runtime.

If we do not have a language version mapping for the runtime, use
"latest" so all its language features are available until we add
the version mapping.

Fix #166 - Support for C# 11