Add host option to disable AssemblyLoadContext
A host can disable the use of AssemblyLoadContext by returning
a value of "true" or "1" when GetHostOption is called with the
option name "DisableAssemblyLoadContext".

This is an escape hatch for hosts that have issues with the ALC
behavior, allowing them to opt out until the underlying issues
can be fixed.