Give host ResolveAssemblyReference priority over reference paths
The list of assembly files referenced by the template may contain
reference assemblies, which will fail to load. Letting the host attempt
to resolve the assembly first gives it an opportunity to resolve runtime
assemblies.

Ideally we would have a robust mechanism for resolving runtime
assemblies bue this will have to serve as a stopgap.